backend/requirements.txt
# Core dependencies
anthropic>=0.40.0
openai>=1.0.0
pymupdf>=1.24.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
tqdm>=4.66.0

# FastAPI & Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Utilities
aiofiles>=23.2.0
pydantic[email]>=2.0.0

# Natural Language Processing (per post-processing anti-AI)
spacy>=3.7.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Image Processing
Pillow>=10.0.0

# Testing
pytest>=7.4.0
httpx>=0.25.0